Changed how wrapping the MetricRegistry works so that Dropwizard Metrics listeners (e.g. JMX reporter) can still function correctly.
Instead of making the adapter override the getter functions for the various metric types we hold a reference to a MetricRegistry and just keep its state in sync with our MetricRegistry by registering/removing metrics as needed but using the dot name (graphite). This makes it so that when Metrics are registered with the Dropwizard MetricRegistry they can still go through the whole MetricRegistryListener chain so reporters such as the JmxReporter will still work and scheduled reports will view the valid metrics (just with the dotted name equivelant). This isn't the most efficient thing to have to keep essentially two maps of Metrics but we only have to deal with this on register/remove, possibly Dropwizard Metrics 4 will be more extensible in this regard (or just have tags out of the box).