Refactor request name handling
Previously we had to provide a getRequestName callback to a RequestTrace
so that the request name can be determined any time. This was required for
the feature stagemonitor.profiler.callStackEveryXRequestsToGroup because
we had to know the name of the request before measurement start. This lead
to overly complex code where we had to somehow get the name before the
measurement starts. Now the request name can be set at any time and can
be influenced by the the user by calling RequestMonitor.getRequest()
.setName(String). It also is much easier to add support for more technologies
like other jax-rs implementations.