Rename anonymize user name to pseudonymize
The difference between pseudonymisation and anonymisation is that
you could never ever track back to whom the data belong if they are anonymized.
With a pseudonym, you could if you had the "key".
Think about a clinical trial for example: Peter is number 09090923. Most researchers
will only handle the number. In case there is need to contact Peter, the head of
department or a third party will have the list of pseudonyms.