Convert (almost) all SPIs to abstract classes
Also (almost) always use a single parameter object for methods
This makes it possible to evolve the SPIs without further breaking changes
See also http://blog.jooq.org/2015/05/21/do-not-make-this-mistake-when-developing-an-spi/

This introduces breaking changes to almost all SPIs which I'm sorry for.
The good news is that the SPIs are much more stable now.