Initial support for prometheus
There are no stagemonitor configuration options to set the
reporting interval and the server yet so you have to add a bit
of code to actually report the metrics.

The reason is that in prometheus, it is not recommended to
push metrics to the server but rather the server should be
configured to pull the metrics. For that to work, you have
to manually register the prometheus MetricsServlet.

See javadoc for StagemonitorPrometheusCollector for more
information

(closes #157)