Convert dotted tag names in nested objects
The problem is that Elasticsearch its not possible to have a property names containing dots.

Even though recent versions of Elasticsearch automatically convert dots in field names into
nested objects it makes sense to do this within stagemonitor so that the json representation
is the same when fetching spans from Elasticsearch and when serializing for the in browser widget.

It also makes no sense to serialize tags like this:
key: theTagKey
value: theTagValue
Because we want to be able to perform aggregations on certain fields, for example to get a list
of the top 10 usernames.