Fix memory leak by not storing Span in SpanContextInformation
Previously, there was a cyclic reference graph between
SpanContextInformation and Span which prevented the Span which
is the key in a WeakConcurrentMap<String, SpanContextInformation>
to be eligible for garbage collection.

Thus, the WeakConcurrentMap grew indefinitely in size with each
new Span.

Additionally, remove the Span from the WeakConcurrentMap after
it is finish()ed.