Add probabilistic sampling options
A default option configures the sampling probability of all spans.
Another option lets you configure a divergent probability for a
specific type of spans like jdbc queries.

The probabilistic sampling can be combined with the rate limiting
sampling so that under normal load, a specified percentage of
spans are reported but at higher loads, the backend won't be
overloaded with spans.