Remove metric aggregation
Breaking change

I don't expect this to be used by anyone
it adds overhead and has very little value.

A more thought through version of this
will come in the future.

Idea: KPI dashboard which can be used to
compare the performance of different
versions/measurement sessions.
For example create one timer to track the
response time for all requests per type.
Configure it in a way that it it does not
account for recency.