Kibana field mapping update improvement
stagemonitor was not able to update the field mapping in Kibana if it was modified (e.g. if the user has refreshed the field mapping manually)
The proposed solution merges the stagemonitor configuration with the current kibana configuration.

The comparison is done on a field-by-field, the strategy looks like this:

* If the kibana index mapping does not yet exist, it is created with the stagemonitor specified default values.
* If the kibana index mapping is somehow broken (fields are not of the expected type) then the merge is aborted.
* If the kibana index mapping exists, field mappings are merged with the following strategy:
  * If a field is only defined in Kibana, it is not modified.
  * If a field exists only in the stagemonitor mapping, it is taken from there.
  * If a field exists in the Kibana mapping and in stagemonitor, it is updated with the stagemonitor version.

(fixes #296)