Sample once per trace
Previously, the sampling decision has been made for each span individually,
but that lead to incomplete and fragmented traces.
Now the sampling decision is only made for root spans so that a trace as a
whole is either completely sampled or not sampled.

One exception is the sampling per individual operation type.
For example you can say I want only 10% of my jdbc spans to be sampled.
This decision is made regardless of the global (per-trace) decision.

An alternative would have been to have a layered sampling like sample
10% of all traces and of those, only sample 50% of the jdbc spans.
This quickly gets quite difficult, especially when you want to see how
many "insertNewUser" statements have been executed within the last hour.
In the example above, if the dashboard says 5 statements within the last
hour, its actually 5/0.5/0.1=100.

The advantage of the current implementation is that you can more easily
tell the effective sampling rate.
If you say jdbc: 0.1 its 10% no matter of the general sampling rate